Fixed lint config

+++ src/bot/scenes/datePickerScene.bot.ts
@@ -6,33 +6,33 @@ import { createInlineKeyboardWithDates } from '../components/inlineKeyboardDateP
 export const datePickerScene = new Scene<BotCustomContext>(BotSceneNameList.DATE_PICKER_SCENE);
 
 datePickerScene.step(async (ctx) => {
-    const datesKeyboard = createInlineKeyboardWithDates();
+	const datesKeyboard = createInlineKeyboardWithDates();
 
-    const keyboardMessage = await ctx.reply('Выберите дату', {
-        reply_markup: datesKeyboard,
-    });
+	const keyboardMessage = await ctx.reply('Выберите дату', {
+		reply_markup: datesKeyboard,
+	});
 
-    ctx.session.keyboardMessageId = keyboardMessage.message_id;
-    ctx.session.chatId = keyboardMessage.chat.id;
+	ctx.session.keyboardMessageId = keyboardMessage.message_id;
+	ctx.session.chatId = keyboardMessage.chat.id;
 
-    ctx.scene.resume();
+	ctx.scene.resume();
 });
 
 datePickerScene.wait('chooseDate').on('callback_query:data', async (ctx) => {
-    await ctx.answerCallbackQuery();
-    const choice = ctx.callbackQuery.data;
-    console.log(choice);
+	await ctx.answerCallbackQuery();
+	const choice = ctx.callbackQuery.data;
+	if (ctx.callbackQuery.message?.message_id === ctx.session.keyboardMessageId){
 
-    ctx.session.tripRequestFilter.date_of_journey = choice;
+		ctx.session.tripRequestFilter.date_of_journey = choice;
 
-    const router = await sceneRouter(ctx);
-    ctx.api.deleteMessage(ctx.session.chatId!, ctx.session.keyboardMessageId!);
-    const nextScene = router.next();
-    if (nextScene === null) {
-        ctx.scene.exit();
-    }
-    else {
-        ctx.scene.enter(nextScene);
-    }
-
+		const router = await sceneRouter(ctx);
+		ctx.api.deleteMessage(ctx.session.chatId!, ctx.session.keyboardMessageId!);
+		const nextScene = router.next();
+		if (nextScene === null) {
+			ctx.scene.exit();
+		}
+		else {
+			ctx.scene.enter(nextScene);
+		}
+	}
 });
